controller and view
Implementation of the Cell and Grid classes has been transferred to
logic.py. The model will contain functions for updating and displaying
cells as well as functionality for starting, stopping, etc.
Controller.py contains methods for creating buttons and has a repeater
to continuously run the simulation. View.py creates the window,
buttons, and canvas.