mouse click for populating/unpopulating cells
User can click on cells in the simulation grid to flip the state of the
board (cell becomes populated if unpopulated and vice versa). This
action is allowed when the simulation is both running and stopped. Note
that users need to click on the board to set up the initial arrangement
of populated cells for practical purposes.